feat: add success guards to non-generic IApiResponse
- Add EnsureSuccessStatusCodeAsync/EnsureSuccessfulAsync extensions for
  the non-generic IApiResponse, mirroring the IApiResponse<T> guards and
  returning the response on success or throwing the captured error
- Reuse the shared GetError helper to surface the captured
  ApiException/ApiRequestException, or a descriptive fallback when none
  was captured
- Home the ensure-success guard tests in a dedicated feature file
- Record the new public surface and document the guards in the README
