+++ src/tests/Refit.Tests/ApiResponseTests.cs
@@ -69,76 +69,6 @@ public sealed class ApiResponseTests
         await Assert.That(ensured).IsSameReferenceAs(response);
     }
 
-    /// <summary>Verifies ensure-success on the interface throws the captured error for an unsuccessful response.</summary>
-    /// <returns>A task representing the asynchronous test.</returns>
-    [Test]
-    public async Task UnsuccessfulResponseEnsureSuccessThrowsCapturedError()
-    {
-        using var request = new HttpRequestMessage(HttpMethod.Get, ExampleUri);
-        using var httpResponse = new HttpResponseMessage(HttpStatusCode.BadRequest) { RequestMessage = request };
-        var error = await ApiException.Create(request, HttpMethod.Get, httpResponse, new());
-        using var response = new ApiResponse<string>(httpResponse, null, new(), error);
-
-        await Assert.That(response.HasContent).IsFalse();
-
-        IApiResponse<string> asInterface = response;
-        await Assert
-            .That(async () => await asInterface.EnsureSuccessStatusCodeAsync())
-            .ThrowsExactly<ApiException>();
-    }
-
-    /// <summary>Verifies EnsureSuccessfulAsync returns the response on success and throws the error on failure.</summary>
-    /// <returns>A task representing the asynchronous test.</returns>
-    [Test]
-    public async Task EnsureSuccessfulAsyncReturnsOnSuccessAndThrowsOnFailure()
-    {
-        using var request = new HttpRequestMessage(HttpMethod.Get, ExampleUri);
-        using var okResponse = new HttpResponseMessage(HttpStatusCode.OK) { RequestMessage = request };
-        using var success = new ApiResponse<string>(okResponse, "body", new());
-
-        IApiResponse<string> okInterface = success;
-        await Assert.That(await okInterface.EnsureSuccessfulAsync()).IsSameReferenceAs(success);
-
-        using var badResponse = new HttpResponseMessage(HttpStatusCode.BadRequest) { RequestMessage = request };
-        var error = await ApiException.Create(request, HttpMethod.Get, badResponse, new());
-        using var failure = new ApiResponse<string>(badResponse, null, new(), error);
-
-        IApiResponse<string> failInterface = failure;
-        await Assert
-            .That(async () => await failInterface.EnsureSuccessfulAsync())
-            .ThrowsExactly<ApiException>();
-    }
-
-    /// <summary>Verifies the success-guard extensions reject a null response.</summary>
-    /// <returns>A task representing the asynchronous test.</returns>
-    [Test]
-    public async Task EnsureSuccessExtensionsRejectNullResponse()
-    {
-        const IApiResponse<string> nullResponse = null!;
-
-        await Assert
-            .That(static async () => await nullResponse.EnsureSuccessStatusCodeAsync())
-            .ThrowsExactly<ArgumentNullException>();
-        await Assert
-            .That(static async () => await nullResponse.EnsureSuccessfulAsync())
-            .ThrowsExactly<ArgumentNullException>();
-    }
-
-    /// <summary>Verifies an unsuccessful response without a captured error throws a descriptive fallback.</summary>
-    /// <returns>A task representing the asynchronous test.</returns>
-    [Test]
-    public async Task EnsureSuccessThrowsFallbackWhenNoErrorCaptured()
-    {
-        using var request = new HttpRequestMessage(HttpMethod.Get, ExampleUri);
-        using var badResponse = new HttpResponseMessage(HttpStatusCode.BadRequest) { RequestMessage = request };
-        using var response = new ApiResponse<string>(badResponse, null, new());
-
-        IApiResponse<string> asInterface = response;
-        await Assert
-            .That(async () => await asInterface.EnsureSuccessStatusCodeAsync())
-            .ThrowsExactly<InvalidOperationException>();
-    }
-
     /// <summary>Verifies public constructors reject missing response data needed by the wrapper.</summary>
     /// <returns>A task representing the asynchronous test.</returns>
     [Test]
